+++ src/resources/views/layouts/appshell3/_apps.blade.php
@@ -1,0 +1,10 @@
+<nav id="nav-sidebar-apps">
+    <nav class="nav-apps-primary">
+        <a href="{{ $appshell->url }}"><img src="{{ $appshell->logoUri ?: '/img/appshell3.svg' }}" class="logo" alt="{{ $appshell->name }}" title="{{ $appshell->name }} Home" /></a>
+        <a href="#" class="nav-app-selected" title="CRM"><i class="fa-light fa-people-group"></i></a>
+    </nav>
+
+    <nav class="nav-apps-secondary">
+        <a href="#"><img src="https://www.gravatar.com/avatar/dd0c7094dfe8ec286241d0fba329e083.jpg?s=100&d=retro" class="avatar avatar-30"></a>
+    </nav>
+</nav>