+++ src/resources/views/layouts/appshell3/_header.blade.php
@@ -1,0 +1,10 @@
+<header class="page-header">
+    <div class="page-header-titlebar">
+        <h1 class="page-title">@yield('title')</h1>
+        <div class="page-actionbar">
+            @yield('page-actions')
+        </div>
+    </div>
+
+    @include('appshell::layouts.appshell3._breadcrumbs')
+</header>