+++ src/resources/views/layouts/appshell3/_sidemenu.blade.php
@@ -1,0 +1,14 @@
+<nav id="nav-sidebar-menu">
+    <h2 class="nav-menu-title">CRM</h2>
+    <nav class="nav-menu">
+        <div class="nav-menu-item nav-menu-group">
+            <span class="nav-menu-group-title">Customers</span>
+            <a href="#">Active</a>
+            <a href="#">Prospects</a>
+            <a href="#">Slipping away</a>
+            <a href="#">Inactive</a>
+            <a href="#">All</a>
+        </div>
+        <a href="#" class="nav-menu-item">Deals</a>
+    </nav>
+</nav>