+++ src/resources/views/user/show.blade.php
@@ -4,6 +4,35 @@
     {{ __('Viewing') }} {{ $user->name }}
 @stop
 
+@section('page-actions')
+    @can('edit users')
+        <x-appshell::button variant="outline-primary" size="sm" href="{{ route('appshell.user.edit', $user) }}">
+            {{ __('Edit user') }}
+        </x-appshell::button>
+    @endcan
+
+    @can('delete users')
+        @if(Auth::id() == $user->id)
+            <x-appshell::button tag="button" variant="outline-danger" size="sm" disabled="disabled" href="{{ route('appshell.user.edit', $user) }}" title="{{ __('It would be quite unhealthy to delete yourself, so you can\'t') }}">
+                {{ __('Delete user') }}
+            </x-appshell::button>
+        @else
+            {!! Form::open([
+                    'route' => ['appshell.user.destroy', $user],
+                    'method' => 'DELETE',
+                    'data-confirmation-text' => __('Are you sure to delete poor :name?', ['name' => $user->name]),
+                    'class' => "inline-block"
+                    ])
+            !!}
+
+            <x-appshell::button tag="button" variant="outline-danger" type="submit" size="sm">
+                {{ __('Delete user') }}
+            </x-appshell::button>
+            {!! Form::close() !!}
+        @endif
+    @endcan
+@stop
+
 @section('content')
 
     <div class="card-deck mb-3">
@@ -62,35 +91,5 @@
         @yield('widgets')
 
     </div>
-
-
-@component(theme_widget('group'))
-    @can('edit users')
-        <a href="{{ route('appshell.user.edit', $user) }}" class="btn btn-outline-primary">{{ __('Edit user') }}</a>
-    @endcan
-
-    @can('delete users')
-        @if(Auth::user()->id == $user->id)
-            <button class="btn btn-outline-danger float-right" disabled="disabled"
-                    title="{{ __("It would be quite unhealthy to delete yourself, so you can't") }}">
-                {{ __('Delete user') }}
-            </button>
-        @else
-            {!! Form::open([
-                    'route' => ['appshell.user.destroy', $user],
-                    'method' => 'DELETE',
-                    'data-confirmation-text' => __('Are you sure to delete poor :name?', ['name' => $user->name]),
-                    'class' => "float-right"
-                    ])
-            !!}
-
-            <button class="btn btn-outline-danger">
-                {{ __('Delete user') }}
-            </button>
-
-            {!! Form::close() !!}
-        @endif
-    @endcan
-@endcomponent
 
 @stop