+++ src/resources/views/widgets/breadcrumbs.blade.php
@@ -1,24 +1,13 @@
-<nav class="breadcrumb">
-    @forelse ($breadcrumbs as $breadcrumb)
-        @if ($breadcrumb->url && !$loop->last)
-            <a class="breadcrumb-item"
-               href="{{ $breadcrumb->url }}">@isset($breadcrumb->icon)
-                    {!! icon($breadcrumb->icon) !!}
-                @endisset{{ $breadcrumb->title }}
-            </a>
-        @else
-            <span class="breadcrumb-item active">@isset($breadcrumb->icon)
-                    {!! icon($breadcrumb->icon) !!}
-                @endisset{{ $breadcrumb->title }}
-            </span>
-        @endif
-    @empty
-        <span class="breadcrumb-item active">
-            &nbsp;
-        </span>
-    @endforelse
-
-    <span class="breadcrumb-menu">
-        @yield('breadcrumb-menu')
-    </span>
+<nav aria-label="breadcrumb">
+    <ol class="breadcrumb">
+        @forelse ($breadcrumbs as $breadcrumb)
+            @if ($breadcrumb->url && !$loop->last)
+                <li class="breadcrumb-item"><a href="{{ $breadcrumb->url }}">@isset($breadcrumb->icon){!! icon($breadcrumb->icon) !!}@endisset{{ $breadcrumb->title }}</a></li>
+            @else
+                <li class="breadcrumb-item active">@isset($breadcrumb->icon){!! icon($breadcrumb->icon) !!}@endisset{{ $breadcrumb->title }}</li>
+            @endif
+        @empty
+            <li class="breadcrumb-item active">&nbsp;</li>
+        @endforelse
+    </ol>
 </nav>