+++ tests/Unit/BadgesWidgetTest.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
 namespace Konekt\AppShell\Tests\Unit;
 
 use Konekt\AppShell\Tests\TestCase;
-use Konekt\AppShell\Theme\AppShellTheme;
+use Konekt\AppShell\Theme\AppShell3Theme;
 use Konekt\AppShell\Widgets\Badges;
 
 class BadgesWidgetTest extends TestCase
@@ -23,7 +23,7 @@ class BadgesWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_default_text_for_empty_lists()
     {
-        $badges = Badges::create(new AppShellTheme(), ['empty' => ['text' => 'Nothing here', 'color' => 'dark']]);
+        $badges = Badges::create(new AppShell3Theme(), ['empty' => ['text' => 'Nothing here', 'color' => 'dark']]);
         $html = trim($badges->render([]));
         $this->assertStringContainsString('badge-dark', $html);
         $this->assertStringContainsString('Nothing here</', $html);