+++ tests/Unit/EnumIconTest.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
 namespace Konekt\AppShell\Tests\Unit;
 
 use Konekt\AppShell\Tests\TestCase;
-use Konekt\AppShell\Theme\AppShellTheme;
+use Konekt\AppShell\Theme\AppShell3Theme;
 use Konekt\AppShell\Widgets\EnumIcon;
 use Konekt\Customer\Models\Customer;
 use Konekt\Customer\Models\CustomerType;
@@ -25,7 +25,7 @@ class EnumIconTest extends TestCase
     /** @test */
     public function it_can_render_if_the_model_is_the_enum_itself()
     {
-        $icon = EnumIcon::create(new AppShellTheme(), ['value' => '$model']);
+        $icon = EnumIcon::create(new AppShell3Theme(), ['value' => '$model']);
         $output = trim($icon->render(CustomerType::INDIVIDUAL()));
 
         $this->assertStringContainsString('<i ', $output);
@@ -35,7 +35,7 @@ class EnumIconTest extends TestCase
     /** @test */
     public function it_can_render_if_the_model_is_an_enum_field_of_the_model()
     {
-        $icon = EnumIcon::create(new AppShellTheme(), ['value' => '$model.type']);
+        $icon = EnumIcon::create(new AppShell3Theme(), ['value' => '$model.type']);
         $customer = new Customer(['type' => CustomerType::ORGANIZATION]);
         $output = trim($icon->render($customer));